# faster/efficient `EmoteExtractor` and `Helpers`