@neon-sunset 2023-07-02
Banger. Out of curiousity, what are the numbers before/after? You can add a new benchmark if you want to https://github.com/TwitchLib/TwitchLib.Client/tree/dev/TwitchLib.Client.Benchmark